rm legacy and move models
