+++ services/web/server/src/simcore_service_webserver/login/_models.py
@@ -1,6 +1,6 @@
 from collections.abc import Callable
 from datetime import datetime
-from typing import Literal
+from typing import Literal, TypedDict
 
 from models_library.users import UserID
 from pydantic import BaseModel, ConfigDict, SecretStr, ValidationInfo
@@ -10,6 +10,19 @@ from .constants import MSG_PASSWORD_MISMATCH
 ActionLiteralStr = Literal[
     "REGISTRATION", "INVITATION", "RESET_PASSWORD", "CHANGE_EMAIL"
 ]
+
+
+class BaseConfirmationTokenDict(TypedDict):
+    code: str
+    action: ActionLiteralStr
+
+
+class ConfirmationTokenDict(BaseConfirmationTokenDict):
+    # SEE packages/postgres-database/src/simcore_postgres_database/models/confirmations.py
+    user_id: int
+    created_at: datetime
+    # SEE handlers_confirmation.py::email_confirmation to determine what type is associated to each action
+    data: str | None
 
 
 class Confirmation(BaseModel):
